refactor(frames): merge Left/Right into FramedLeft/FramedRight
Phase B6 of frame unification. Replace the four Left/Right components with
two config-driven ones:
- FramedLeft: icons/sizes from getFrameConfig(style).left; the one structural
  fork is the new-file button (Gnome ButtonWindow vs Windows ButtonTool).
- FramedRight: shared menu/close handlers; the structural fork is Gnome's
  grouped .menu-btn/.close-group vs Windows' flat layout.
Per-frame .left/.controls geometry is scoped by [data-frame]; both layouts
kept verbatim → pixels identical.

Also fix a stale GNOME_CONFIG.right.menu size (24 -> 16) to match the
rendered GnomeFrame menu button. DOM signature byte-identical for both
styles (the menu icon is sampled, so this is verified lossless).
