+++ src/renderer/Panel/frames/GnomeFrame/Panel.svelte
@@ -1,16 +1,16 @@
 <script lang="ts">
   import { getFrameStyleVars } from "Utils/Render/frameTheme";
+  import FramedLeft from "../FramedLeft.svelte";
+  import FramedRight from "../FramedRight.svelte";
   import FramedTabs from "../FramedTabs.svelte";
-  import Left from "./Left.svelte";
-  import Right from "./Right.svelte";
 
   let { zoom = 1 } = $props<{ zoom?: number }>();
 </script>
 
 <div id="panel" data-frame="gnome" style="zoom: {zoom}; {getFrameStyleVars('gnome')}">
-  <Left />
+  <FramedLeft style="gnome" />
   <FramedTabs style="gnome" />
-  <Right />
+  <FramedRight style="gnome" />
 </div>
 
 <style>
